eval: isolate model-matrix probe from environment hook mocks
